Clean definition of Direct Stream Access APIs
The remaining and remainingUpTo APIs were thought to provide
a direct access to the stream for the case of pollers. We have not
taken care of the specifics of these APIs and this change aims to revisit
or redefine the behavior.

- remaining should be able to load from a checkpoint and return the entries represented by it.
- remaining should be able to load from a checkpoint and the regular stream in a single call.
- remaining should not load from a checkpoint if the stream is seeked (pointer is moved) as
we might return duplicates.
- If loading from a checkpoint the stream's global pointer should point to the address it represents
in the regular stream and not the actual address of the checkpoint entry itself.

Pending: because remaining/remainingUpTo can return entries that come from the checkpoint stream or
the regular stream (transactional or not), these APIs should return a List<SMREntry> instead of
List<ILogData> as it is not expected that a consumer deals with the complexities of a checkpoint/regular stream.