Sequencer Reconfiguration Corruption Fixes
Prior to this patch, the global tail, stream tails and stream bit
sets were retrieved and computed from the first node in the last
segment. This assumption is incorrect and leads to the possibility
of reconfiguring the sequencer with state that doesn't reflect the
log. In consequence, leading to inconsistent reads, data loss and
cluster instability.

The patch introduces multiple fixes to compute the sequencer state
from the right head node (i.e. the head node of all segments) and
forces the computation part and reconfiguration to happen on the
same epoch.