Sequencer Reconfiguration Corruption Fixes (#2722)
Prior to this patch, the global tail, stream tails and stream bit
sets were retrieved and computed from the first node in the last
segment. This assumption is incorrect because it can be a node
that is being rebuilt (via state transfer), or in some edge cases
it is not possible to compute the sequencer state from a single node.
This leads to the possibility of computing a view of the log that
is incomplete during reconfiguration of the sequencer. In
consequence, leading to inconsistent reads, data loss and
cluster instability.

This patch introduces multiple fixes to compute the sequencer's
state from all the segments (possibly aggregating views from different
nodes) and forcing the view computation and the reconfiguration to
happen on the same epoch.