Introduce Compatibility Vector for RPCs
We introduce the compatibility vectors for communicating the information about features that are supported by the sender and receiver.

We represent each feature as an enum value in a Protobuf file and set its corresponding enum number's index to 1 in the BitSet.

The BitSet values will be the same across both the client and the server of a particular version.