Change LogData to use Java Supplier over guava (#2971)
The current implementation uses guava based Supplier, it throws IncompatibleClassChangeError when a java based supplier is passed in its place as observed corfudb-cloud repo test cases. Though both are interchangeable the auto-conversion is not happening in some cases. The time() method expects the Supplier to be a Java based one. This commit changes it to use Java one over guava one.