Stream Polling Delay Period Fix
This patch removes random sleeps between polling and replaces it
with a fixed sleep duration, which makes stream polling follow a
fixed delay schedule. This also reduces the load being generated
by the pollers and makes performance more predictable.

Pre-this patch when the user specifies a poll period of 50ms, the
scheduler will use a random sleep value between [0, 50), this means
that the poller will run every 24.5ms (i.e., expected value of
Random(50)) on average, in essence doubling the frequency, which
is not expected by the user.