Sequencer Bootstrap Inconsistency
A query to a nonexistent stream returns a StreamAddressSpace with
a default trim mark, that is, -1. But if the stream has one write
and the logunit and sequencer restart, then the new sequencer's
view will be bootstrapped with a StreamAddressSpace that has the
trim mark initialized to -6. This change changes the behavior such
that the trim mark is always consistent.