Advance Stream Tags on Checkpoint (#3071)
Advance stream tags on checkpoint for UFO tables. Since stream tags
represent streams and are not checkpointed, the tag's stream trim
mark can't be used to detect trim gaps since they are always
uninitialized. This patch advances all streams that represent
a tag for a particular stream on checkpoint, that way stream listeners
that listen on tags can efficiently detect whether a trim is detected
or not.