StreamingManager Rewrite (#3047)
This patch re-implements the streaming implementation to improve
efficiency and reduce latency. This PR improves the previous
implementation by reducing the number of tunable parameters, batching
and coalescing poll requests for multiple streams, this allows for the
removal of many to stream threads that busy-poll the sequencer. As a
consequence, the total number of threads is reduced, significantly
reduces the sequencer polling load and is able to discover changes
faster.