Verify Returned Objects References (#3085)
This patch adds a test that verifies that the internal Map.Entry
is not leaked to the API callers. Since these containers can be
reused internally in the map, after returning it to the caller
the value can change which can cause consistency/correctness
issues where different threads observe inconsistent view of the
key/value for the same version.

Related PR3068: Fixed a correctness issue to the aforementioned
problem and this PR adds a test to prevent future regressions.