Remove seperate codacy reporter action
Earlier we used to run unit tests all over again for reporting
code coverage results on all the pull requests.
This commit reduces this step to re-use the existing unit tests.

We used one action for the code coverage reporting on pull requests
and on pushes to master.
This commit splits that action to report code coverage of -
- Pull requests after the existing unit tests job.
- Pushes on master during the existing artifacts deploy job.

Note: It changes artifact publish step mvn build to
      run tests that were skipped earlier. But thats
      a far lesser overhead compared to the existing set up.