Persist EPOCH in LogEntry
'Epoch' is one of several fields captured by the ILogData metadata map.
We have a bug where the epoch (metadata) is never persisted as part of LogEntry.
Therefore, entries read from disk (no longer available in the server's cache) do not have the epoch and are
returned to client's as epoch -1 (for instance in the case of StreamListeners).