Exclude 'clearing' EMPTY streams to replicate (#3132)
Clearing 'empty' streams on snapshot sync start can cause "apparent" data loss issues in
subsequent swithovers. This is because a 'clear' enforces an update to a stream that has
not been opened by upper layers. If the stream has not been opened it will not be checkpointed.
On a switchover, when attempting to read this stream (as it is intended for replication) we can
hit a TrimmedException.