CorfuStoreBrowser: Fix printing of all schemas in registry table
Because registry table's TableDescriptors uses Any protobuf
type the browser fails to print the registry table which
makes it hard to figure out what the schemas are in any given setup.
This fix avoids using the JsonPrinter for the
TableDescriptor since there is a known issue.
CorfuStoreBrowser: fix up loader to re-use table entry
Minor fixes and comments from reviewers