commit() returns >0 address if no streams are touched in txn
If the streams accessed in a transaction are empty
commit() would return -1 as the commit() address
which makes it impossible to use as a stream
re-subscription point.
So if this corner case is detected, return the address from
transaction's snapshot already obtained from the sequencer
Also force a read at that address to materialize it.