TableRegistry: Use GitRepo version to detect Protobuf file changes
Currently if a protobuf file has changed we do not capture the
latest changes in the ProtobufDescriptor table.
This can cause issues if the key's type has changed since the
compactor depends on the key's type to make its KeyDynamicProtobuf
But since the protobuf map comparison is not deterministic
we version the protobuf file descriptor and on first open
after upgrade we just update the schema.