Bug fix: when CP entries are streamed, set global address as CP log start address
Originally, when stream.streamUpTo(v) process CP entries, the global address
of the CP entry is set to the address of the CP entry itself, not the snapshot
address at which the CP is taken. This messes up the version mechanism in MVO.

Refer to CheckpointSmokeTest::MultiCheckpointWriter2Test.
Failure: In MultiVersionObject::syncStreamUnsafe, stream.streamUpTo(92) returns a
CP entry at addr 94 and put the object@94 to MVOCache. When reading the table again,
getVersionedObject(92) hit StaleObjectVersionException because MVOCache says the
lowest version is 94.