Atomic apply of updates during snapshot sync (#3253)
Currently, snapshot sync updates received for a stream are not applied atomically
on the Sink. First, a stream is cleared. In a subsequent transaction,
the updates(SMR entries) are applied one-by-one, each in a separate transaction.

This increases the window during which applications observe inconsistent data.

This window can be minimized by atomically clearing + applying all(if possible)
updates. If the size of the transaction exceeds CorfuRuntime's MAX_WRITE_SIZE,
the updates can be applied in batches.