Disable Caching and Add Support For Floor Queries (#3336)
This patch exposes a new floor API for StreamAddressSpace. This core
operation will allow MVCC to efficiently determine the latest version that
is visible from a particular timestamp.  This patch also disables caching
within the Roaring64NavigableMap backing the address space in order
to allow for concurrent optimistic read operations on the data structure.