Fix flaky test CorfuReplicationClusterConfigIT.testActiveLockRelease()
This test did not account for the delay in the asynchronous processing
of a lock release/expiry.  This made it to fail intermittently.

The fix reduces the overall wait time required in the test and also
accounts for the async event processing delay by waiting for the
Replication Status to change to STOPPED.