@paxos: Change how the prepare rank is incremented
When the prepare rank is strictly monotonically increasing, under
certain conditions -- close to synchronized failure detector
execution on all nodes, along with network delay -- the cluster can
enter a live-lock type of state where the majority of clients will
attempt to fill a Paxos slot/register and continue to out-rank one
another during the prepare phase of Paxos.

In order to avoid this scenario, the rank should be incremented by some
random amount in order to ensure faster convergence.