Test Fix for Redundant SnapshotSync Updates in Multi-Source -> Single Sink Topology. (#3502)
LR currently has a limitation that the Snapshot Writers on Sink cannot
differentiate between the replicated streams received from their own vs other
sessions.  So if multiple writers find a stream in the global list of streams to
be replicated, each will apply updates from the shadow stream of that stream,
even though no data was received for it from the writer's own session.
The current PR is a workaround in the test to prevent it from
failing while the real fix is under discussion.