Improve The Failure Detection Time
This patch improves the response time of CorfuRuntime::fetchLayout
by trying to discover the latest Layout with a broadcast instead
of randomly sampling nodes. Randomizing the fetch is problematic
because the selection process is sequential and it can choose an
unresponsive node, and since the fault detector relies on that
method, it can add more latency to the fault detection period.