Independent runtime for compactor 410 (#3615)
* Handle exception thrown during init of CompactorService

When initializing optional variables in CompactorService, if there's any exception,
it's not handled and hence results in the CompactorService thread being killed. This
leads to the runOrchestrator not being invoked and so compactor never gets triggered.
This commit fixes the above issue.

* Create new runtime for CompactorService

The CompactorService needs to register a systemDownHandler where the
runtime is reinitialized on an unrecoverable exception. Since the CompactorService
uses the same runtime as the other services part of the management server, we
need an independent runtime for CompactorService.

* Handle concurrency on systemDownHanlder in CompactorService

* Make systemDownHandlerForCompactor non-recursive and deadlock free

* Fix flaky CompactorServiceIT test