[Backport] Fix sequencer cache eviction policy (#3010) (#3626)
* Fix sequencer cache eviction policy (#3010)

The sequencer cache implements a sliding window to help with cache
eviction, which incorrectly assumes that conflict keys are unique.
This can lead to a memory leak when the same keys are written over
and over again. This patch also fixes a bug where the cache evicts
partial conflicts keys for a particular timestamp. This can lead to
incorrect resolutions and can cause data loss.

* Use latest Ubuntu version for GitHub actions (#3614)

---------

Co-authored-by: Zachary Frenette <66146392+zfrenette@users.noreply.github.com>