Upgrade tests must use the Source Status Listener to avoid intermittent failures.
Listening for expected number of updates on the Replication Status table
and SnapshotSync Plugin on the Sink can lead to intermittent failures if the
Snapshot Sync is restarted by the Source due to a flaky network or out-of-order
message delivery.  Verification of snapshot sync completion by listening
on the Source replication status of COMPLETED avoids test failures in
the above scenarios and is reliable.

Also reduced the overall test time by removing waits on the Sink Status
table.