Fix Intermittent Failures in Upgrade tests (#3634)
Listening for expected number of updates on the Replication Status table
and SnapshotSync Plugin on the Sink can lead to intermittent failures if the
Snapshot Sync is restarted by the Source due to a flaky network or out-of-order
message delivery.  Verification of snapshot sync completion by listening
to is_dataConsitent=true on the Sink Replication Status table avoids test failures in
the above scenarios and is reliable.