Fix the flakiness in LogReplicationFSMTest.testSnapshotSender. (#3638)
testSnapshotSender relies on observing the state transition to a
non-terminal state(InSnapshotSync).  It relies on an Observable to
deliver this update and then directly validates the FSM state.  But the
FSM could have transitioned to the next state, i.e.,
WaitForSnapshotApply or InLogEntrySync by the time the Observable
invokes the update() callback.  Change the test to rely on the terminal
state, i.e., InLogEntrySyncState to make it reliable.

Also removed the dependence on the semaphore for this test as it is much simpler
to wait for the expected state directly.