Update Corfu-411 with Master TOT (#3653)
* Reset trimmed table on reopen to avoid throwing TrimmedException (#3588)

* Reset trimmed table on reopen to avoid throwing TrimmedException

On upgrade scenarios, it is possible that some tables are new to a version
and when they are accessed in some corner cases, a TrimmedException is thrown.
This fix on reopening a table, mimicks the compactor by appending a HOLE to
the empty stream, thus avoiding the TrimmedException.

* Do resetTrimmedTable within a transaction

* Add metrics server (#3599)

* expose addTypeToMap for rolling upgrades (#3601)

This allows new code to open old schemas without
actually registering the old types.

* Reset object state on exceptions during sync (#3602)

This patch fixes a bug where clients can observe stale or incomplete
data when an exception occurs on some code paths during a sync.
By resetting the state of the object and its SMR stream, subsequent
reads will be served from an accurate view of the object.

* Bump netty version from 4.1.77 to 4.1.89 (#3565)

* Bump netty version from 4.1.77 to 4.1.89

Bump netty versions to fix [CVE-2022-41881](https://nvd.nist.gov/vuln/detail/CVE-2022-41881)

Compatibility: https://github.com/grpc/grpc-java/blob/master/SECURITY.md#netty

grpc.version 1.54.0
netty.version 4.1.89.Final
netty.tcnative.version 2.0.56.Final

* fix GRPC sample code after grpc upgrade

* netty 4.1.89.Final

reactor-netty-http doesn't have 4.1.87.Final
in their released versions.
reactor-netty-http 1.1.3 has 4.1.89.Final.

---------

Co-authored-by: Shama Hegde <shama358@gmail.com>

* Create new runtime for CompactorService (#3611)

* Create new runtime for CompactorService

The CompactorService needs to register a systemDownHandler where the
runtime is reinitialized on an unrecoverable exception. Since the CompactorService
uses the same runtime as the other services part of the management server, we
need an independent runtime for CompactorService.

* Handle concurrency on systemDownHanlder in CompactorService

* Make systemDownHandlerForCompactor non-recursive and deadlock free

* Fix flaky CompactorServiceIT test

* Adding the missing max-write-size to CorfuServerCmdLine. (#3617)

The msx-write-size has been renamed to max-replication-write-size

* Increase AddNode pingRetries to 10 to avoid failing too fast (#3621)

Existing pingRetries and pingInterval in AddNode give a 20-30s window.
This windows is too short for a corfu server restart. Increase the
pingRetries from 3 to 10 to allow ~60s timeout window.

* proto 3.21.7 (#3618)

* Fix Flaky Tests in NettyCommTest, CompileProxyTest, DeltaStreamTest (#3624)

* Fix Flaky Tests in NettyCommTest, CompileProxyTest

- NettyCommTest:
Make bootstrapServer() wait for the server to be up and running before checking for ping messages.

Move all combinations of ciphers and keystore args into separate methods to enable threads clean up after each test.

- CompileProxyTest:
Increase timeout for executeScheduled as it takes more than 1 second to  execute the scheduled the task.

* Fix DeltaStreamTest.concurrencyTest() timeout

Increase `done` latch's wait timeout to 10 seconds

* Fix compactor tests (#3632)

- Fix flaky compactor test - validateLivenessNonLeaderTest
- Remove sleeps in CompactorServiceUnitTest

* Update BoundedMap PreConditions (#3639)

* Avoid opening compactor metadata tables on every shouldTrigger call (#3600)

When shouldTrigger initializes DistributedCheckpointerHelper on every call,
it opens all the compactor metadata tables leading to excessive openTable logs.
Hence, reuse distributedCheckpointerHelper from CompactorService

* Fix the flakiness in LogReplicationFSMTest.testSnapshotSender. (#3638)

testSnapshotSender relies on observing the state transition to a
non-terminal state(InSnapshotSync).  It relies on an Observable to
deliver this update and then directly validates the FSM state.  But the
FSM could have transitioned to the next state, i.e.,
WaitForSnapshotApply or InLogEntrySync by the time the Observable
invokes the update() callback.  Change the test to rely on the terminal
state, i.e., InLogEntrySyncState to make it reliable.

Also removed the dependence on the semaphore for this test as it is much simpler
to wait for the expected state directly.

* registerTable needs to retry more times before giving up (#3642)

Right after upgrade if a large number of tables are
opened in parallel, the chances of collisions increase
many fold since many tables may share the same protobuf schemas
under the hood.

* Add 'ServerAddress' Compactor Config (#3636)

Make compactor take a ServerAddress config
from YML file whose value can be replaced
by corfu-server init script

---------

Co-authored-by: Sravanthi A <ashokkumars@vmware.com>
Co-authored-by: Viacheslav <314292+xnull@users.noreply.github.com>
Co-authored-by: Sundar Sridharan <sundar.sridharan@gmail.com>
Co-authored-by: Zachary Frenette <66146392+zfrenette@users.noreply.github.com>
Co-authored-by: Shama Hegde <shama358@gmail.com>
Co-authored-by: Jie (George) Lu <jielu@vmware.com>
Co-authored-by: Pankti Majmudar <pmajmudar@vmware.com>