Do not set batch read size to 1 when cache is disabled or 0 (#3693)
When AddressSpaceView readCache is 0, do not reset batch read size to 1 as it will make state transfer very do 1 address per batch partition instead of 100.

This causes CPU and network overhead due to the increased number of threads and network calls across nodes. This further causes cluster instability and delay in join node operations as it starts the state transfer.