Introduce Read-Only Scoped Transactions (#3866)
Introduce a new transaction type (called ScopedTxn) whose lifespan is
defined by the lifespan of the ScopenTxn instance, and is not tied the
thread local storage.

When defining a scoped transaction, the client has to specify a set of
tables he or she is interested in during the initialization, which will
effectively create a snapshot transaction. The underlying tables are not
tied to MVO cache and are not eligible for eviction.