Backpressure Slow Connections
Some pathological network conditions can cause asymmetric latencies
where a client keeps submitting requests quickly, but isn't able
to consume the response fast enough. Causing it to retry while the
server is still trying to respond to earlier requests. This latency
mismatch between the forward path and backwards path can consume
an inordinate amount of memory from the direct memory pool causing
allocation failures. Essentially leaking memory until the channel
is closed, or when normal network conditions are restored.

This patch introduces a simple back pressure mechanism blocking
for some time when a channel is not writable, then closing the
connection if the outbound channel is not depleted fast enough.